Grouped pins in Arduino Pro Mini symbol and removed pins that don't get pads.
Having power in/out through the top/bottom is more conventional for
schematics.

Eliminating the pins that don't get pads avoids errors in board layout.

Putting the D13/LED_BUILTIN on the right is motivated by the desire to
have outputs on the right.  Digital and analog pins can be used as inputs
or outputs, but I probably should have considered putting the analog ones
on the left, since they're specialized for analog input.